Function reference

This page contains the complete documentation of Repo functions. Between I have inserted some comments and explanations of the underlying philosophy, as well as some usage notes.

Construction

Repo objects can be constructed either from an existing 1-dimensional array or by specifying the element type and number. In both cases the number of classes must be specified.

Repos.Repo — Method

Repo{T}(N::Int,C::Int) where T

Create a repository that organizes N elements of type T into C classes.

Example

julia> using Repos

julia> r=Repo{String}(50000,4)
Repository of 50000 objects in 4 classes

Repos.Repo — Method

Repo(a::Array,C::Int)

Create a repository that organizes an existing array into C classes.

The repository will not copy the array, so the memory is only allocated once and any changes you make in the array will be reflected in the repository.

Example

julia> using Repos

julia> a=["red","green","blue"];

julia> r=Repo(a,2)
Repository of 3 objects in 2 classes

julia> a[1]="cyan";

julia> print_repo(r)
Repository of 2 classes
  Class 1
    1 - 1: cyan
    2 - 2: green
    3 - 3: blue
  Class 2

Repo is optimized to deal with very large number of elements. The runtime of most functions should be independent of the number of elements and linear in the number of classes.

Once the repository has been created the number of elements and classes cannot be changed. If you are unsure of how many elements you need, make the Repo large enough and assign unused elements to one class, so one class basically becomes the storage space for unused elements.

Class functions

The central concept of Repo is that elements stored in the Repo can be categorized into different classes, where each class is identified by a numerical value. You can set and retrieve the class of an element by the following two functions:

Repos.getclass — Method

getclass(repo::Repo{T},id::Int)

Return the class number of element id in repo.

Example

julia> using Repos

julia> r=Repo(["red","green","blue"],2)
Repository of 3 objects in 2 classes

julia> getclass(r,1)
1

Repos.setclass! — Method

setclass!(repo::Repo{T},id::Int,cls::Int) where T

Change the class of item number id in repo to cls.

Example

julia> using Repos

julia> r=Repo(["red","green","blue"],2)
Repository of 3 objects in 2 classes

julia> setclass!(r,1,2)

julia> print_repo(r)
Repository of 2 classes
  Class 1
    1 - 3: blue
    2 - 2: green
  Class 2
    1 - 1: red

An interesting function is the class() function this function returns an interface object of type RepoClass. This interface acts as a portal to access the elements in the respective class. It will always remain up to date (No data is actually copied into the interface object).

Repos.class — Method

 class(repo::Repo{T},c::Int)

Return an interface to class c in repository repo.

The returned RepoClass structure is an interface to the object of the respective class in the repository. The RepoClass will stay up to date when the repository is updated.

Example

julia> using Repos

julia> a=["Golden Delicious","Granny Smith","Red Delicious","Valencia"]
4-element Array{String,1}:
 "Golden Delicious"
 "Granny Smith"
 "Red Delicious"
 "Valencia"

julia> r=Repo(a,2)
Repository of 4 objects in 2 classes

julia> apples=class(r,1)
Class of 4 objects

julia> setclass!(r,4,2)

julia> a[3]="Braeburn"
"Braeburn"

julia> print_repo(apples)
Class of 3 objects
    1 - 1: Golden Delicious
    2 - 2: Granny Smith
    3 - 3: Braeburn

ID and Length

There are two ways to identify an element in the Repo. First, you can identify an element by it's ID number. This number will always refer to the same element. If you created the Repo from an array, the ID will be identical to the elements number in the array.

Second, you can identify an element by its class and position in the class. This is useful for example for iterating through a class or picking a random element from the class. However, in contrast to ID the position of an element in a class can change when setclass!() is called. If you say, pick using class and position and want to remember which element you picked for later use, you should store it's ID which can be determined by the following functions:

Repos.repo_id — Method

repo_id(repo::Repo{T},c::Int,pos::Int)
repo_id(repo::Repo{T},pos::Int)
repo_id(cls::RepoClass{T},pos::Int)

Get the permanent id of an item in the repository identified by current position or current class and and current position.

In contrast to the position, which can change when setclass is called, the id of an item never changes.

Example

julia> using Repos

julia> r=Repo([1:10...],3)
Repository of 10 objects in 3 classes

julia> repo_id(r,1)
1

julia> setclass!(r,1,3)

julia> repo_id(r,1)
10

julia> repo_id(r,3,1)
1

To determine the size of the repository or the elements in a certain class use the length() function.

Example

julia> using Repos

julia> villain=Repo(["Dracula","Frankenstein","Hyde","Catwoman"],2)
Repository of 4 objects in 2 classes

julia> setclass!(villain,4,2)

julia> victorian=class(villain,1)
Class of 3 objects

julia> modern=class(villain,2)
Class of 1 objects

julia> length(villain)
4

julia> length(victorian)
3

julia> length(modern)
1

Element access

You can access element with "[]" like you would in a vector. The elements can be identified either by their id or by class and position. (see the notes on position, above)

Example

julia> using Repos

julia> numbers=Repo([1:1000...],2)
Repository of 1000 objects in 2 classes

julia> setclass!(numbers,123,2)

julia> numbers[1]
1

julia> numbers[123]
123

julia> numbers[2,1]
123

julia> numbers[2,1]=-5
-5

julia> numbers[123]
-5

julia> numbers[1]=-7
-7

julia> numbers[1,1]
-7

If you created a Repo from an array you can also change the repo elements by changing the original array (Repo does not hardcopy the array at creation.)

Example

julia> using Repos

julia> cities=["Kuala Lumpur","Moscow","San Francisco","London","Rio de Janeiro"]
5-element Array{String,1}:
 "Kuala Lumpur"
 "Moscow"
 "San Francisco"
 "London"
 "Rio de Janeiro"

julia> repo=Repo(cities,2)
Repository of 5 objects in 2 classes

julia> capitals=class(repo,1)
Class of 5 objects

julia> setclass!(repo,3,2)

julia> setclass!(repo,5,2)

julia> print_repo(repo)
Repository of 2 classes
  Class 1
    1 - 1: Kuala Lumpur
    2 - 2: Moscow
    3 - 4: London
  Class 2
    1 - 5: Rio de Janeiro
    2 - 3: San Francisco

julia> cities[5] = "Bristol"
"Bristol"

julia> repo[5]
"Bristol"

julia> repo[2,1]
"Bristol"

Iteration

You can iterate over either the Repo or a RepoClass using standard "for" iteration.

Example

julia> using Repos

julia> rock=Repo(["Granite","Diorite","Gabbro","Andesite","Basalt"],2)
Repository of 5 objects in 2 classes

julia> setclass!(rock,3,2)

julia> setclass!(rock,5,2)

julia> mafic=class(rock,2)
Class of 2 objects

julia> for x in rock println(x) end
Granite
Diorite
Gabbro
Andesite
Basalt

julia> for x in mafic println(x) end
Basalt
Gabbro

When iterating over the whole Repo the elements are presented in order of element id. When iterating over a class the elements are presented in the order of position withinthe class. So the order can change whenever setclass!() is called.

In the case where functions in the for loop change the class of elements the iteration will still run over all objects that were in the respective class at the start of the loop. So every object that is in the respective class when the iteration starts will be iterated over exactly once.

If different behavior is desired iterate over the elements using the length() and [].

Printing

For debugging purposes an overview of a Repo or RepoClass object can be printed.

Repos.print_repo — Method

print_repo(repo::Repo{T},nax::Int=10)
print_repo(cls::RepoClass{T},max::Int=10)
print_repo(io::IO,repo::Repo{T},nax::Int=10)
print_repo(io::IO,cls::RepoClass{T},max::Int=10)

Print a summary of the contents of the repository repo or the repository class cls to REPL or another IO device.

The output lists the classes and shows the first and last entries in every class. For each entry the current position in the class, the id of the entry, and the content of the entry are printed.

An optional parameter max can be specified to control the number of entries that are shown in every repository class. If this parameter is set to less than 2 all entries are shown. (Warning: For large repositories this can take a while)

Example

julia> using Repos

julia> r=Repo(["A","B","C","D","E","F","G"],2)
Repository of 7 objects in 2 classes

julia> setclass!(r,1,2)

julia> setclass!(r,5,2)

julia> vowels=class(r,2)
Class of 2 objects

julia> print_repo(vowels)
Class of 2 objects
    1 - 5: E
    2 - 1: A

julia> r=Repo([1:10000...],2)
Repository of 10000 objects in 2 classes

julia> setclass!(r,1337,2)

julia> print_repo(r,4)
Repository of 2 classes
  Class 1
    1 - 1: 1
    2 - 2: 2
    ...
    9998 - 9998: 9998
    9999 - 9999: 9999
  Class 2
    1 - 1337: 1337